create Tree functor
A balanced tree implemention, which is similar to
AVL trees, but with slightly relaxed balancing that
tests show to be faster than either standard AVL trees
or Red-Black trees.

Also, conversion to streams is provided, allowing for
faster iteration than the conversion to lists provided
with the SML/NJ library's Red-Black trees. Inexact search
is also possible.